+++ crates/action/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+moon_contract = { path = "../contract" }
 moon_utils = { path = "../utils" }
 clap = { version = "4.0.10", features = ["derive"] }
 serde = { workspace = true }
