+++ crates/action/src/lib.rs
@@ -1,5 +1,7 @@
 mod action;
 mod context;
+mod node;
 
 pub use action::*;
 pub use context::*;
+pub use node::*;
