+++ crates/emitter/Cargo.toml
@@ -4,8 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+moon_action = { path = "../action" }
+moon_contract = { path = "../contract" }
 moon_error = { path = "../error" }
+moon_project = { path = "../project" }
+moon_task = { path = "../task" }
 moon_utils = { path = "../utils" }
 moon_workspace = { path = "../workspace" }
 async-trait = "0.1.57"
+serde = { workspace = true }
 tokio = { workspace = true }
