+++ crates/emitter/src/subscriber.rs
@@ -3,7 +3,7 @@ use moon_error::MoonError;
 use moon_workspace::Workspace;
 
 #[async_trait::async_trait]
-pub trait Subscriber {
+pub trait Subscriber: Send + Sync {
     async fn on_emit<'e>(
         &mut self,
         event: &Event<'e>,
