+++ crates/runner/src/lib.rs
@@ -1,12 +1,9 @@
 mod actions;
 mod dep_graph;
-mod emitter;
 mod errors;
-mod node;
 mod runner;
 mod subscribers;
 
 pub use dep_graph::*;
 pub use errors::*;
-pub use node::ActionNode;
 pub use runner::*;
